+++ dwarfeng-ftp-api/src/test/java/com/dwarfeng/ftp/api/integration/configuration/ServiceExceptionMapperConfiguration.java
@@ -1,0 +1,27 @@
+package com.dwarfeng.ftp.api.integration.configuration;
+
+import com.dwarfeng.subgrade.impl.exception.MapServiceExceptionMapper;
+import com.dwarfeng.subgrade.sdk.exception.ServiceExceptionHelper;
+import com.dwarfeng.subgrade.stack.exception.ServiceException;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import java.util.Map;
+
+/**
+ * 服务异常映射配置。
+ *
+ * @author DwArFeng
+ * @since 2.0.0
+ */
+@Configuration
+public class ServiceExceptionMapperConfiguration {
+
+    @Bean
+    public MapServiceExceptionMapper mapServiceExceptionMapper() {
+        Map<Class<? extends Exception>, ServiceException.Code> destination = ServiceExceptionHelper.putDefaultDestination(null);
+        destination = com.dwarfeng.ftp.sdk.util.ServiceExceptionHelper.putDefaultDestination(destination);
+        destination = com.dwarfeng.springtelqos.sdk.util.ServiceExceptionHelper.putDefaultDestination(destination);
+        return new MapServiceExceptionMapper(destination, com.dwarfeng.subgrade.sdk.exception.ServiceExceptionCodes.UNDEFINED);
+    }
+}